// src/app/event-data.type.ts
import { ResisterUser } from './resistered-user.type';
export interface EventData {
  card_tags?: string[];
  cover_picture?: string;
  currency?: string;
  end_time?: number;
  event_category?: string;
  event_start_time?: number;
  event_end_time?: number;
  event_subcategory?: string;
  fees?: number;
  registered_users?: ResisterUser;
  has_started?: boolean;
  highlight_event?: boolean;
  id?: number;
  is_college_specific?: boolean;
  mobile_cover_picture?: string;
  name?: string;
  orderable_key?: string;
  registration_start_time?: number;
  registration_end_time?: number;
  registration_status?: string;
  seats_filled?: number;
  seats_left?: number;
  short_desc?: string;
  slug?: string;
  start_time?: number;
  user_already_registered?: string;
  venue?: string;
}